+++ tests/ImboUnitTest/EventListener/StatsAccessTest.php
@@ -73,98 +73,62 @@ class StatsAccessTest extends ListenerTests {
         return array(
             'IPv4 in whitelist' => array(
                 '127.0.0.1',        // IP
-                array('127.0.0.1'), // Whitelist
-                array(),            // Blacklist
+                array('127.0.0.1'), // Allow
                 true                // Access?
             ),
             'IPv4 not in whitelist' => array(
                 '127.0.0.2',
                 array('127.0.0.1'),
-                array(),
-                false
-            ),
-            'IPv4 not in blacklist' => array(
-                '127.0.0.2',
-                array(),
-                array('127.0.0.1'),
-                true
-            ),
-            'IPv4 in blacklist' => array(
-                '127.0.0.1',
-                array(),
-                array('127.0.0.1'),
-                false
-            ),
-            'IPv4 in whitelist with both filters populated' => array(
-                '127.0.0.1',
-                array('127.0.0.1', '127.0.0.2'),
-                array('127.0.0.3', '127.0.0.4'),
-                true
-            ),
-            'IPv4 in blacklist with both filters populated' => array(
-                '127.0.0.3',
-                array('127.0.0.1', '127.0.0.2'),
-                array('127.0.0.3', '127.0.0.4'),
-                false
-            ),
-            'IPv4 in both lists with both filters populated' => array(
-                '127.0.0.2',
-                array('127.0.0.1', '127.0.0.2'),
-                array('127.0.0.2', '127.0.0.3'),
-                false
-            ),
-            'IPv4 in no lists with both filters populated' => array(
-                '127.0.0.5',
-                array('127.0.0.1', '127.0.0.2'),
-                array('127.0.0.3', '127.0.0.4'),
                 false
             ),
             'IPv4 in whitelist range' => array(
                 '192.168.1.10',
                 array('192.168.1.0/24'),
-                array(),
                 true
-            ),
-            'IPv4 in blacklist range' => array(
-                '192.168.1.10',
-                array(),
-                array('192.168.1.0/24'),
-                false
             ),
             'IPv4 outside of whitelist range' => array(
                 '192.168.1.64',
                 array('192.168.1.32/27'),
-                array(),
                 false
             ),
             'IPv6 in whitelist (in short format)' => array(
                 '2a00:1b60:1011:0000:0000:0000:0000:1338',
                 array('2a00:1b60:1011::1338'),
-                array(),
                 true
             ),
             'IPv6 in whitelist (in full format)' => array(
                 '2a00:1b60:1011:0000:0000:0000:0000:1338',
                 array('2a00:1b60:1011:0000:0000:0000:0000:1338'),
-                array(),
                 true
             ),
             'IPv6 in whitelist range' => array(
                 '2001:0db8:0000:0000:0000:0000:0000:0000',
                 array('2001:db8::/48'),
-                array(),
                 true
             ),
             'IPv6 outside of whitelist range' => array(
                 '2001:0db9:0000:0000:0000:0000:0000:0000',
                 array('2001:db8::/48'),
-                array(),
                 false
             ),
             'IPv6 in whitelist (in short format in both fields)' => array(
                 '2a00:1b60:1011::1338',
                 array('2a00:1b60:1011::1338'),
-                array(),
+                true
+            ),
+            'Blaclisted IPv4 client and both types in allow' => array(
+                '1.2.3.4',
+                array('127.0.0.1', '::1'),
+                false
+            ),
+            'Whitelitsed IPv6 client and both types in allow' => array(
+                '::1',
+                array('127.0.0.1', '::1'),
+                true
+            ),
+            'Wildcard allows all clients' => array(
+                '::1',
+                array('*'),
                 true
             ),
         );
@@ -173,14 +137,13 @@ class StatsAccessTest extends ListenerTests {
     /**
      * @dataProvider getFilterData
      */
-    public function testCanUseDifferentFilters($clientIp, $whitelist, $blacklist, $hasAccess) {
+    public function testCanUseDifferentFilters($clientIp, $allow, $hasAccess) {
         $this->request->expects($this->once())
                       ->method('getClientIp')
                       ->will($this->returnValue($clientIp));
 
         $listener = new StatsAccess(array(
-            'whitelist' => $whitelist,
-            'blacklist' => $blacklist,
+            'allow' => $allow,
         ));
 
         if (!$hasAccess) {